requirements.txt
pydantic
pydantic-settings
sqlalchemy
python-dotenv
requests
ccxt
pandas
pandas-ta
openai
yfinance
alpaca-py
alpha_vantage
